Some modifications make it possible to read user variables and do arithmentic in the NtC command.
More substantially:
I intrudouced the overrideAtomicProperty command. This changes the specified atomic property in the atomInfoVector of every BiopolymerClass in the BiopolymerClassContainer, At the moment only the atomicNumber property can be changed. Note that negative values are allowed. This is useful for density map fitting which uses the MDFF potential, meaning that the force is proportional to the atomic number. This means that etting the backbone atomic numbers to negative values helps keep the double helix centered in long tubes of low-resolution density. For example:

overrideAtomicProperty P   atomicNumber -15.