created new sphericalSpiral command. Syntax:  sphericalSpiral <spherical radius, in nm> <DNA inter-helix distance (suggest 2.0), in nm> <helix start (nearest to north pole) theta, rads> <helix start phi, rads>
This creates a spherical spiral which the user should adjust to follow the DNA within one winding shell of a virus.

To visualize the result, spiral.pdb is written. This contains MG2+ atoms which follow the spiral. They are spaced 3.4Å apart.

Also, there is produced a commands.spiral.dat . This contains tetherToGround commands, in readAtStage blocks.

You can put the contents of the commands.spiral.dat in your commands file. At each stage, one of the tethers will be placed. I have an example file where there is one base pair produced and fitted at each stage. Best to see that example for further explanation.