added the new parameter useNACappingHydroxyls.
If in the command file you issue:

useNACappingHydroxyls 0

Then the 5' nucleic acid residue will have a 5' phosphate.

If you do NOT set this parameter, it defaults to 1, meaning that the phosphorus will be replaced by an H5T.

This was done to enable us to create DNA by generating bases one at a time.