The trajectory mmCIF files are now written and closed after each trajectory is computed, reading and adding to them by the next trajectory.
This has the advantage that all mmCIF trajectories do no longer need to be kept in memory until the last iteration is done, but it required a more reading and writing (I assume this is not a big issue).
Also, now the mmCIF trajectory files contain only remarks relating the trajectories written into them, instead of all previous remarks.