After Michal Tykac's updates to enable mmCIF, most things worked great. However I later found that the chain prefix was not working .  This is a prefix, specified as the second parameter of the loadSequenceFromPdb command. The prefix is prepended to all chain IDs obtained from the structure file in question. this is useful when more than one source pdb file is used, and (as often happens) they both use the same chain ID or IDs.  "A" is the most popular chain ID, for instance, so any two PDB files are likely to both have an "A" chain.
To be more specific, the fault was that the chains with the prefix had the right sequence but no structure. So it was the matchCoordinates function that was not handling prefixes correctly.

The problem was that in BiopolymerClass.cpp, matchCoordinates and loadSequencesFromPdb are two functions which both separately open the input PDB file(s). The purpose is to match coordinates and read sequences, respectively. loadSequencesFromPdb was correctly applying the prefix. However matchCoordinates was not.

The fix was to use the more sophisticated of the two file readers, which was the one in matchCoordinates. This one supports .gz suffixes on file names, and also has more error trapping. I factored this out into a separate function, not a member of BiopolymerClass or its Container.  Both matchCoordinates and loadSequencesFromPdb then called this one function. the function loads the PdbStructure into the pdbStructureMap, so if it is called a second time (in matchCoordinates) it first checks the Map and only generates PdbStructure if the latter is not found in the Map
 .  As part of this it was necessary to take pdbStructureMap (a member of BiopolymerClassContainer) as a parameer in the mentioned BiopolymerClass members matchCoordinates.

I tested this with a variant of commands.protein-morphing.dat , modified to use the prefix, and it worked great.