create new parameters for alignmentForces: matchScore mismatchScore gapPenalty scoringScheme
scoringScheme takes Simple or Blosum62. The first three only apply for scoringScheme Simple. Now you can have more control over your alignment, for instance if you wish to make "hard" alignment which forces the full length of the two sequences to be aligned 1:1 with no concern for the actual sequence.